+++ packages/import-html/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @templatical/import-html
 
+## 0.17.0
+
+### Patch Changes
+
+- @templatical/types@0.17.0
+
+## 0.16.5
+
+### Patch Changes
+
+- @templatical/types@0.16.5
+
+## 0.16.4
+
+### Patch Changes
+
+- Updated dependencies [1801876]
+  - @templatical/types@0.16.4
+
 ## 0.16.3
 
 ### Patch Changes